+++ src/test/java/site/nomoreparties/stellarburgers/OrderAPITest.java
@@ -4,12 +4,11 @@ import io.qameta.allure.Step;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.Response;
 import org.apache.http.HttpStatus;
-import org.hamcrest.Matchers;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import site.nomoreparties.stellarburgers.model.Ingredient;
+import site.nomoreparties.stellarburgers.model.response.Ingredient;
 import site.nomoreparties.stellarburgers.model.Burger;
 import site.nomoreparties.stellarburgers.model.response.IngredientResponse;
 import site.nomoreparties.stellarburgers.model.response.MessageResponse;
