+++ src/test/java/site/nomoreparties/stellarburgers/model/response/IngredientResponse.java
@@ -3,7 +3,7 @@ package site.nomoreparties.stellarburgers.model.response;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import site.nomoreparties.stellarburgers.model.Ingredient;
+
 import java.util.List;
 
 @Getter
