+++ src/test/java/site/nomoreparties/stellarburgers/model/response/OrderResponse.java
@@ -3,7 +3,6 @@ package site.nomoreparties.stellarburgers.model.response;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import site.nomoreparties.stellarburgers.model.Ingredient;
 import site.nomoreparties.stellarburgers.model.response.orderdetail.OrderDetail;
 
 @Getter
