+++ src/test/java/site/nomoreparties/stellarburgers/model/response/orderdetail/OrderDetail.java
@@ -4,7 +4,6 @@ import com.google.gson.annotations.SerializedName;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import site.nomoreparties.stellarburgers.model.Ingredient;
 
 import java.util.List;
